Removing a larger set of classes.
