+++ java/ham_fisted/MutList.java
@@ -1,6 +1,5 @@
 package ham_fisted;
 
-import static ham_fisted.BitmapTrieCommon.*;
 import static ham_fisted.ChunkedList.*;
 
 import java.util.List;
